Release 1.0.6

+++ backend/super-bi/src/main/java/net/aopacloud/superbi/service/impl/SysPasswordServiceImpl.java
@@ -62,7 +62,7 @@ public class SysPasswordServiceImpl implements SysPasswordService {
             retryCount = 0;
         }
 
-        if (retryCount >= Integer.valueOf(maxRetryCount).intValue()) {
+        if (retryCount >= maxRetryCount) {
             LoginEvent loginEvent = LoginEvent.of(username,
                     Constants.LOGIN_FAIL,
                     LocaleMessages.getMessage("user.password.retry.limit.exceed", maxRetryCount, lockTime),
